refactor: update esbuild configuration paths and add esbuild dependency
