+++ Common/package.json
@@ -84,6 +84,7 @@
         "crypto-js": "^4.2.0",
         "dotenv": "^16.4.4",
         "ejs": "^3.1.10",
+        "esbuild": "^0.25.5",
         "express": "^4.21.1",
         "formik": "^2.4.6",
         "history": "^5.3.0",
